backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
numpy==1.26.4
scikit-learn==1.3.2
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
sqlalchemy==2.0.34
passlib==1.7.4
python-jose==3.3.0
opencv-python==4.8.1.78
joblib==1.3.2
scipy==1.11.4
torch==2.2.2
torchvision==0.17.2
boto3==1.34.131
GitPython==3.1.43
pytest==8.2.0